+++ lib/process-data-projects.ts
@@ -2,7 +2,7 @@ import { ProjectType } from "@/interfaces/_common-data-types";
 
 import { NewProjectData, ProjectData, ProjectDocPopulated } from "@/interfaces/Project";
 
-import { fileDocuments_toData } from "./process-data-files";
+import { fileDocuments_toData } from "./process-data-files-mongodb";
 import { tagDocuments_toData } from "./process-data-tags";
 import { processMarkdown } from "./process-markdown";
